+++ mon/MonDescWaveform.hh
@@ -22,8 +22,8 @@ namespace Pds {
     float xup() const;
 
   private:
-    unsigned short _nbins;
-    unsigned short _unused;
+    unsigned _nbins;
+    unsigned _unused;
     float _xlow;
     float _xup;
   };
